refactor(studio): rename to studio and flatten component layout
Rename the shell app apps/midscene-studio -> apps/studio (package name,
docs/plan filename and contents) and flatten the renderer component tree
by removing the incut-area-all-in-one directory layer so every shell
component lives directly under src/renderer/components.

- Move IncutAreaAllInOne, MainContent, Playground, Sidebar, globals.css
  up into src/renderer/components and drop the redundant barrel index.ts
- Rename the asset helper to asset-urls.ts to avoid a file-vs-directory
  collision with the sibling assets/ folder
- Update App.tsx and IncutAreaAllInOne.tsx imports to the new sibling
  paths, and tighten the midscene-icon.png relative path in asset-urls
- Refresh docs/plan to reference apps/studio and pnpm --filter studio

Validated with pnpm --filter studio build and pnpm run lint.
