--- apps/midscene-studio/src/renderer/components/incut-area-all-in-one/index.ts
@@ -1,1 +1,0 @@
-export { default } from './IncutAreaAllInOne';
