+++ apps/studio/assets/midscene-icon.png
