+++ apps/studio/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "midscene-studio",
+  "name": "studio",
   "private": true,
   "version": "0.1.0",
   "type": "module",
