+++ apps/studio/scripts/sync-static-assets.mjs
