+++ apps/studio/src/main/index.ts
