+++ apps/studio/src/renderer/App.css
