+++ apps/studio/src/renderer/components/MainContent.tsx
@@ -1,4 +1,4 @@
-import { incutAssetUrls } from '../assets';
+import { incutAssetUrls } from './asset-urls';
 
 export default function MainContent() {
   return (
