+++ apps/studio/src/renderer/components/assets/main-chat.png
