+++ apps/studio/src/renderer/components/assets/main-device.png
