+++ apps/studio/src/renderer/components/assets/sidebar-computer.png
