+++ apps/studio/src/renderer/components/assets/sidebar-overview.png
