+++ apps/studio/src/renderer/index.tsx
