+++ apps/studio/src/shared/electron-contract.ts
