+++ docs/plan/studio-phase-1.md
@@ -2,8 +2,8 @@
 
 ## Goal
 
-Create a new `apps/midscene-studio` shell application powered by `Rsbuild`
-and Electron.
+Create a new `apps/studio` shell application powered by `Rsbuild` and
+Electron.
 
 ## Scope
 
@@ -13,7 +13,7 @@ and Electron.
   - `main`: Electron main process bundle
   - `preload`: typed IPC bridge for the renderer
 - Ship a light-themed native-feeling two-pane layout
-- Vendor shell assets inside `midscene-studio` so the app does not depend on
+- Vendor shell assets inside `apps/studio` so the app does not depend on
   remote design-export URLs at runtime
 - Keep the runtime layering explicit so phase 2 can plug in real device and
   agent flows without reorganizing the app
@@ -21,7 +21,7 @@ and Electron.
 ## Architecture Decisions
 
 - `src/main`: owns native window lifecycle and OS-specific Electron
-  integration; static assets are synced from `apps/midscene-studio/assets`
+  integration; static assets are synced from `apps/studio/assets`
   into `dist/assets` by `scripts/sync-static-assets.mjs`
 - `src/preload`: the only trusted renderer bridge; business logic stays out
 - `src/renderer`: UI only, no direct Node or Electron imports
@@ -29,5 +29,5 @@ and Electron.
 
 ## Validation
 
-- `pnpm --filter midscene-studio build`
+- `pnpm --filter studio build`
 - `pnpm run lint`
